Move out division code and delete the old `arithmetic.rs`
